requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.31.0
ray>=2.0.0
pyyaml>=6.0.0
python-lsp-server>=1.7.0
python-lsp-jsonrpc>=1.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
jsonschema>=4.18.0
jsonschema>=4.18.0

# Optional: For enhanced features
fastapi>=0.100.0  # For web API examples
uvicorn>=0.20.0   # ASGI server
requests>=2.28.0  # HTTP client
python-dotenv>=1.0.0  # Load .env for CLI backend
python-multipart>=0.0.9  # FastAPI file uploads
python-multipart>=0.0.9  # FastAPI file uploads
